requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
protobuf==5.28.3
scikit-learn==1.5.2
numpy==2.1.3
joblib==1.4.2
